+++ source/renderer/app/components/staking/stake-pools/hooks/useCreateColumns.tsx
@@ -1,94 +1,31 @@
 import React, { useMemo } from 'react';
-import { orderBy } from 'lodash';
 import classNames from 'classnames';
 import BigNumber from 'bignumber.js';
 import moment from 'moment';
 import { FormattedHTMLMessage } from 'react-intl';
+import { Column } from 'react-table';
 import { PopOver } from 'react-polymorph/lib/components/PopOver';
-import { PoolPopOver } from '../widgets/PoolPopOver';
-import { Intl } from '../../../types/i18nTypes';
-
-import styles from './StakePoolsTable.scss';
-import StakePool from '../../../domains/StakePool';
-import { getColorFromRange, getSaturationColor } from '../../../utils/colors';
+import { PoolPopOver } from '../../widgets/PoolPopOver';
+import { Intl } from '../../../../types/i18nTypes';
+import styles from '../StakePoolsTable.scss';
+import StakePool from '../../../../domains/StakePool';
+import {
+  getColorFromRange,
+  getSaturationColor,
+} from '../../../../utils/colors';
 import {
   formattedWalletAmount,
   toFixedUserFormat,
-} from '../../../utils/formatters';
-import { messages } from './StakePoolsTable.messages';
-
-const ascOrder = 'asc';
-const descOrder = 'desc';
-
-export const defaultTableOrdering = {
-  ranking: ascOrder,
-  ticker: ascOrder,
-  saturation: ascOrder,
-  cost: ascOrder,
-  profitMargin: ascOrder,
-  producedBlocks: descOrder,
-  nonMyopicMemberRewards: descOrder,
-  pledge: ascOrder,
-  retiring: ascOrder,
-};
-
-interface UseSortedStakePoolListArgs {
-  stakePoolList: StakePool[];
-  sortBy: string;
-  order: 'asc' | 'desc';
-}
-
-export const useSortedStakePoolList = ({
-  stakePoolList,
-  sortBy,
-  order,
-}: UseSortedStakePoolListArgs) =>
-  useMemo(
-    () =>
-      orderBy(
-        stakePoolList.map((stakePool) => {
-          let calculatedPledge;
-          let calculatedCost;
-          let formattedTicker;
-
-          if (sortBy === 'ticker') {
-            formattedTicker = stakePool.ticker
-              .replace(/[^\w\s]/gi, '')
-              .toLowerCase();
-          }
-
-          if (sortBy === 'pledge') {
-            const formattedPledgeValue = stakePool.pledge.toFixed(2);
-            calculatedPledge = Number(
-              parseFloat(formattedPledgeValue).toFixed(2)
-            );
-          }
-
-          if (sortBy === 'cost') {
-            const formattedCostValue = stakePool.cost.toFixed(2);
-            calculatedCost = Number(parseFloat(formattedCostValue).toFixed(2));
-          }
-
-          return {
-            ...stakePool,
-            calculatedPledge,
-            calculatedCost,
-            formattedTicker,
-          };
-        }),
-        ['formattedTicker', 'calculatedPledge', 'calculatedCost', sortBy],
-        [order, order, order, order]
-      ),
-    [stakePoolList, order, sortBy]
-  );
+} from '../../../../utils/formatters';
+import { messages } from '../StakePoolsTable.messages';
 
 type UseCreateColumnsArgs = {
   currentTheme: string;
   showWithSelectButton?: boolean;
-  onSelect?: (...args: Array<any>) => any;
+  onSelect?: (poolId: string) => void;
   containerClassName: string;
   numberOfRankedStakePools: number;
-  onOpenExternalLink: (...args: Array<any>) => any;
+  onOpenExternalLink: (url: string) => void;
   intl: Intl;
 };
 
@@ -101,7 +38,7 @@ export const useCreateColumns = ({
   containerClassName,
   showWithSelectButton,
 }: UseCreateColumnsArgs) =>
-  useMemo(
+  useMemo<Column<StakePool>[]>(
     () => [
       {
         id: 'ranking',
@@ -120,7 +57,7 @@ export const useCreateColumns = ({
         ),
         accessor: 'ranking',
         Cell: ({ row }) => {
-          const { potentialRewards, ranking }: StakePool = row.original;
+          const { potentialRewards, ranking } = row.original;
           const memberRewards = new BigNumber(potentialRewards);
 
           return (
@@ -143,7 +80,7 @@ export const useCreateColumns = ({
         Header: intl.formatMessage(messages.tableHeaderTicker),
         accessor: 'ticker',
         Cell: ({ row }) => {
-          const stakePool: StakePool = row.original;
+          const stakePool = row.original;
           const color = getColorFromRange(
             stakePool.ranking,
             numberOfRankedStakePools
@@ -180,7 +117,7 @@ export const useCreateColumns = ({
         ),
         accessor: 'saturation',
         Cell: ({ row }) => {
-          const { saturation }: StakePool = row.original;
+          const { saturation } = row.original;
           const progressBarContentClassnames = classNames([
             styles.progressBarContent,
             styles[getSaturationColor(saturation)],
@@ -273,7 +210,7 @@ export const useCreateColumns = ({
         ),
         accessor: 'nonMyopicMemberRewards',
         Cell: ({ row }) => {
-          const stakePool: StakePool = row.original;
+          const stakePool = row.original;
           const memberRewards = new BigNumber(stakePool.potentialRewards);
           const potentialRewards = formattedWalletAmount(memberRewards);
           return potentialRewards;
@@ -292,7 +229,7 @@ export const useCreateColumns = ({
         ),
         accessor: 'pledge',
         Cell: ({ row }) => {
-          const stakePool: StakePool = row.original;
+          const stakePool = row.original;
           const pledge = new BigNumber(stakePool.pledge);
           const pledgeValue = formattedWalletAmount(pledge, false, false);
           return pledgeValue;
@@ -303,7 +240,7 @@ export const useCreateColumns = ({
         Header: intl.formatMessage(messages.tableHeaderRetiring),
         accessor: 'retiring',
         Cell: ({ row }) => {
-          const stakePool: StakePool = row.original;
+          const stakePool = row.original;
           const retirement =
             stakePool.retiring &&
             moment(stakePool.retiring).locale(intl.locale).fromNow(true);